cleaned-up some code, verified stoikov a bit more, got it running fix to fee schedule and place order, ping fix changed timestamps on market data, so timestamp is unix timestamp with dot after second, this is similar to Binance implementation, also remember that algorithms like stoikov refresh on clock tick of one second

+++ hummingbot/core/data_type/order_book_tracker.py
@@ -5,6 +5,7 @@ from collections import deque
 from enum import Enum
 import logging
 import pandas as pd
+from decimal import Decimal
 import re
 from typing import (
     Dict,
@@ -296,8 +297,8 @@ class OrderBookTracker(ABC):
                 order_book.apply_trade(OrderBookTradeEvent(
                     trading_pair=trade_message.trading_pair,
                     timestamp=trade_message.timestamp,
-                    price=float(trade_message.content["price"]),
-                    amount=float(trade_message.content["amount"]),
+                    price=Decimal(trade_message.content["price"]),
+                    amount=Decimal(trade_message.content["amount"]),
                     type=TradeType.SELL if
                     trade_message.content["trade_type"] == float(TradeType.SELL.value) else TradeType.BUY
                 ))
